+++ crates/madara/primitives/chain_config/src/chain_config.rs
@@ -464,7 +464,7 @@ mod tests {
     #[rstest]
     fn test_mainnet_from_yaml() {
         // Change the current directory
-        std::env::set_current_dir("../../../").expect("Failed to change directory");
+        std::env::set_current_dir("../../../../").expect("Failed to change directory");
         let chain_config: ChainConfig =
             ChainConfig::from_yaml(Path::new("configs/presets/mainnet.yaml")).expect("failed to get cfg");
 
@@ -480,8 +480,9 @@ mod tests {
 
         // Check versioned constants
         // Load and parse the JSON file
-        let json_content = fs::read_to_string("crates/primitives/chain_config/resources/versioned_constants_13_0.json")
-            .expect("Failed to read JSON file");
+        let json_content =
+            fs::read_to_string("crates/madara/primitives/chain_config/resources/versioned_constants_13_0.json")
+                .expect("Failed to read JSON file");
         let json: Value = serde_json::from_str(&json_content).expect("Failed to parse JSON");
 
         // Get the VersionedConstants for version 0.13.0