+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 
+- chore: Move crates under a madara subdir
 - chore(nix): resolve flake and direnv compatibility issues
 - fix: Gateway path fix
 - fix: instrumentation code